Refactor: Optimize SchemaKit with modular architecture and simplified exports
